requirements.txt
azureml-sdk
torch-vision
torch
